+++ src/py/flwr/common/record/__init__.py
@@ -17,7 +17,7 @@
 
 from .configsrecord import ConfigsRecord
 from .conversion_utils import array_from_numpy
-from .metricsrecord import MetricsRecord
+from .metricrecord import MetricsRecord
 from .parametersrecord import Array, ParametersRecord
 from .recorddict import RecordDict, RecordSet
