+++ src/py/flwr/common/record/recorddict.py
@@ -23,7 +23,7 @@ from typing import TypeVar, Union, cast
 
 from ..logger import log
 from .configsrecord import ConfigsRecord
-from .metricsrecord import MetricsRecord
+from .metricrecord import MetricsRecord
 from .parametersrecord import ParametersRecord
 from .typeddict import TypedDict
